+++ packages/terra-popup/tests/jest/PopupDefaultExample.jsx
@@ -29,6 +29,8 @@ class PopupDefaultExample extends React.Component {
           isOpen={this.state.open}
           targetRef={this.getButtonNode}
           onRequestClose={this.handleRequestClose}
+          requestFocus={() => {}}
+          releaseFocus={() => {}}
         >
           <p>this is popup content</p>
         </Popup>