+++ packages/terra-site/src/examples/popup/PopupTopLeft.jsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import Popup from 'terra-popup';
+import Button from 'terra-button';
+import ExamplePopupContent from './ExamplePopupContent';
+
+class PopupTopLeft extends React.Component {
+  constructor(props) {
+    super(props);
+    this.handleButtonClick = this.handleButtonClick.bind(this);
+    this.handleRequestClose = this.handleRequestClose.bind(this);
+    this.handleOnChange = this.handleOnChange.bind(this);
+    this.state = { open: false };
+  }
+
+  handleButtonClick() {
+    this.setState({ open: true });
+  }
+
+  handleRequestClose() {
+    this.setState({ open: false });
+  }
+
+  handleOnChange() {
+    this.setState({ open: false });
+  }
+
+  render() {
+    return (
+      <div>
+        <Popup
+          contentAttachment="top left"
+          isArrowDisplayed
+          isOpen={this.state.open}
+          onRequestClose={this.handleRequestClose}
+          targetRef={() => document.getElementById('popup-top-left')}
+        >
+          <ExamplePopupContent onChange={this.handleOnChange} />
+        </Popup>
+        <Button id="popup-top-left" text="Top-Left" onClick={this.handleButtonClick} />
+      </div>
+    );
+  }
+}
+
+export default PopupTopLeft;