+++ packages/terra-modal-manager/CHANGELOG.md
@@ -15,6 +15,9 @@ Unreleased
 ### Fixed
 * Removed unnecessary layout styles (Issue #610)
 
+### Added
+* Added requestFocus and releaseFocus actions to support trapping and releasing focus in the modal manager.
+
 1.0.0 - (June 28, 2017)
 ------------------
 Initial stable release